Makes all read counting dependent on args.verbose
On our HPC Cluster with a shared networked filesystem and 1000+ matam jobs
running at the same time our filesystem was saturated due to the many `wc -l`
subprocesses.

By looking at the `matam_assembly.py` script we realized that all the read
counting done with `wc -l` is only used for logging purposes and does not
actually affect any computation. That's why we decided to make it optional
based on `args.verbose`.